separate out server code
