+++ src/message.h
@@ -3,6 +3,7 @@
  */
 
 #include <stdio.h>
+#include <stdbool.h>
 
 #include "log.h"
